@@ -1,7 +1,6 @@
 import { HttpErrorResponse } from '@angular/common/http'
-import { Observable } from "rxjs/Observable";
+import { Observable } from "rxjs";
 import { ErrorHandler, Injectable, Injector, NgZone } from "@angular/core";
-import 'rxjs/add/observable/throw'
 
 import { NotificationService} from './shared/messages/notification.service'
 import {LoginService} from './security/login/login.service'
