@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {FormGroup, FormBuilder, Validators, AbstractControl} from '@angular/forms'
+import {FormGroup, FormBuilder, FormControl, Validators, AbstractControl} from '@angular/forms'
 import { RadioOption } from "../shared/radio/radio-option.model";
 import { CartItem } from "../restaurant-detail/shopping-cart/cart-item.model";
 import { OrderService } from "../order/order.service";
@@ -7,7 +7,7 @@ import { Order, OrderItem } from "./order.model";
 
 import {Router} from '@angular/router'
 
-import 'rxjs/add/operator/do'
+import {tap} from 'rxjs/operators'
 
 @Component({
   selector: 'mt-order',
@@ -32,15 +32,15 @@ export class OrderComponent implements OnInit {
   constructor(private orderService: OrderService, private router: Router, private formBuilder: FormBuilder) { }
 
   ngOnInit() {
-    this.orderForm = this.formBuilder.group({
-        name: this.formBuilder.control('', [ Validators.required, Validators.minLength(5)]),
-        email: this.formBuilder.control('', [ Validators.required, Validators.pattern(this.emailPattern)]),
-        emailConfirmation: this.formBuilder.control('', [ Validators.required, Validators.pattern(this.emailPattern)]),
-        address: this.formBuilder.control('', [ Validators.required, Validators.minLength(5)]),
-        number: this.formBuilder.control('', [ Validators.required, Validators.pattern(this.numberPattern)]),
-        optionalAddress: this.formBuilder.control(''),
-        paymentOption: this.formBuilder.control('', [ Validators.required])
-    }, {validator: OrderComponent.equalsTo})
+    this.orderForm = new FormGroup({
+        name: new FormControl('', {validators : [Validators.required, Validators.minLength(5)]}),
+        email: new FormControl('', {validators: [ Validators.required, Validators.pattern(this.emailPattern)]}),
+        emailConfirmation: new FormControl('', {validators: [ Validators.required, Validators.pattern(this.emailPattern)]}),
+        address: new FormControl('', {validators: [ Validators.required, Validators.minLength(5)]}),
+        number: new FormControl('', {validators: [ Validators.required, Validators.pattern(this.numberPattern)]}),
+        optionalAddress: new FormControl(''),
+        paymentOption: new FormControl('', {validators: [ Validators.required]})
+    }, {validators: [OrderComponent.equalsTo], updateOn: 'blur'})
   }
 
   static equalsTo(group: AbstractControl): {[key:string]: boolean}
@@ -83,9 +83,9 @@ export class OrderComponent implements OnInit {
     order.orderItems = this.cartItems()
       .map((item:CartItem)=> new OrderItem(item.quantity, item.menuItem.id))
       this.orderService.checkOrder(order)
-        .do((orderId: string) =>{
+        .pipe(tap((orderId: string) =>{
           this.orderId = orderId
-        })
+        }))
         .subscribe((orderId: string) => {
           this.orderService.clear()
           this.router.navigate(['/order-summary'])
