@@ -4,12 +4,10 @@ import { RestaurantsService } from "./restaurants.service";
 import { trigger, state, style, transition, animate} from '@angular/animations'
 import {FormBuilder, FormControl, FormGroup} from '@angular/forms'
 
-import 'rxjs/add/operator/switchMap'
-import 'rxjs/add/operator/debounceTime'
-import 'rxjs/add/operator/distinctUntilChanged'
-import 'rxjs/add/operator/catch'
-import 'rxjs/add/observable/from'
-import {Observable} from 'rxjs/Observable'
+
+import {Observable, from } from 'rxjs'
+
+import {switchMap, tap, debounceTime, distinctUntilChanged, catchError } from 'rxjs/operators'
 
 @Component({
   selector: 'mt-restaurants',
@@ -39,11 +37,14 @@ export class RestaurantsComponent implements OnInit {
     })
 
     this.searchControl.valueChanges
-    .debounceTime(500)
-    .distinctUntilChanged()
-    .switchMap(searchTerm =>
-      this.restaurantService.restaurants(searchTerm)
-      .catch(error => Observable.from([])))
+    .pipe(
+      debounceTime(500),
+      distinctUntilChanged(),
+      switchMap(searchTerm =>
+        this.restaurantService.restaurants(searchTerm)
+        .pipe(catchError(error => from([])))
+      )
+    )
     .subscribe(restaurants => this.restaurants = restaurants)
 
     this.restaurantService.restaurants().subscribe(restaurants => this.restaurants = restaurants)
@@ -51,7 +52,6 @@ export class RestaurantsComponent implements OnInit {
 
   toggleSearch()
   {
-    console.log('teste')
     this.seachBarState = this.seachBarState === 'hidden' ? 'visible': 'hidden'
   }
 
