@@ -1,8 +1,8 @@
 import {Injectable} from '@angular/core'
 import {HttpClient, HttpParams} from '@angular/common/http'
-import {Observable} from 'rxjs/Observable'
-import 'rxjs/add/operator/map'
-import 'rxjs/add/operator/catch'
+
+import {Observable} from 'rxjs'
+import {map, catchError} from 'rxjs/operators'
 
 import {Restaurant} from "./restaurant/restaurant.model"
 import {MEAT_API} from '../app.api'
