@@ -1,5 +1,5 @@
 import { HttpInterceptor, HttpRequest, HttpHandler, HttpEvent } from "@angular/common/http";
-import { Observable } from "rxjs/Observable";
+import { Observable } from "rxjs";
 import { LoginService } from './login/login.service'
 import { Injectable, Injector } from "@angular/core/";
 import { HttpHeaders } from "@angular/common/http";
