@@ -3,10 +3,13 @@ import { Injectable } from "@angular/core";
 import { CartItem } from "../restaurant-detail/shopping-cart/cart-item.model";
 import { ShoppingCartService } from "../restaurant-detail/shopping-cart/shopping-cart.service";
 import { Order, OrderItem } from "./order.model";
-import { Observable } from "rxjs/observable";
+
 import { HttpClient, HttpHeaders } from '@angular/common/http'
-import 'rxjs/add/operator/map'
 import { MEAT_API } from "../app.api";
+
+import { Observable } from "rxjs";
+import { map } from 'rxjs/operators'
+
 
 @Injectable()
 export class OrderService{
@@ -40,7 +43,7 @@ export class OrderService{
 
   checkOrder(order: Order): Observable<String> {
     return this.http.post<Order>(`${MEAT_API}/orders`, order)
-                    .map(orderId => order.id)
+                    .pipe(map(orderId => order.id))
   }
 
 }
