@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import {ActivatedRoute} from '@angular/router'
 import {RestaurantsService} from '../../restaurants/restaurants.service'
 import { MenuItem } from "../menu-item/menu-item.model";
-import { Observable } from "rxjs/observable";
+import { Observable } from "rxjs";
 
 @Component({
   selector: 'mt-menu',
