@@ -17,6 +17,6 @@ import 'core-js/es6/reflect';
 
 import 'core-js/es7/reflect';
 import 'zone.js/dist/zone';
-import 'intl'
-import 'intl/locale-data/jsonp/pt-br.js'
+//import 'intl'
+//import 'intl/locale-data/jsonp/pt-br.js'
 import 'web-animations-js/web-animations.min.js'
